+++ config/swagger-definition.js
@@ -1,0 +1,12 @@
+module.exports = {
+  openapi: '3.0.1',
+  info: {
+    title: 'GROWI API v3',
+    version: '3.0.0',
+  },
+  externalDocs: {
+    description: 'GROWI Docs',
+    url: 'https://docs.growi.org',
+  },
+  basePath: '/api/v3/',
+};